Refactored API
Did some improvements on carrier and shipper API.
Carrier and Shipper are now automatically filled with null and the active user's id respectively when you use POST method.
There is now a many to many relationship between Load and Carrier (currently as Shipper) to register dropped loads.
Personal shipper information is showed instead of the id.